Fixed validation incorrectly passing with whitespace-only user agent (#386)
Currently a request with a whitespace-only user agent header, i.e.
`User-Agent: ` can pass validation. This causes errors downstream in the
`BatchWorker`, which ends up trying to process a message with an empty
user agent.

This fixes the problem with the request validation, which should prevent
these invalid events from reaching the `BatchWorker`. It does not yet
improve the handling in the `BatchWorker` itself.