Fixed BatchWorker getting stuck on unprocessable messages (#387)
The `BatchWorker` currently `nacks` any message that it encounters an
error in the course of processing. However, some errors are
deterministic: if we fail to parse a message the first time, we're going
to fail to parse it in the future as well. Rather than `nacking` the
message, only to retry it again and again, it's better to log the error
and `ack` the message, which will remove it from the queue.

Arguably this should be done via deadlettering, so Pub/Sub will try to
deliver the message X times and then add it to a deadletter queue if all
attempts fail. For now at least, I think it makes sense to `ack` any
message that we know if going to repeatedly fail. Once we get
deadlettering setup properly, we can revisit this behavior.