Updated target in compose.yml to use development build (#423)
The `compose.yml` file didn't specify a build target, so it was
defaulting to the last one (production). The production build doesn't
have all the necessary dependencies for local development, so `yarn dev`
was failing. This sets the target properly, so when running `docker
compose *` commands, it will use the development image instead of the
production image.