Wired up UTM params to processed payload (#429)
ref https://linear.app/ghost/issue/PROD-2561/
- updated various schema; hoisted UTM params from within referrerSource
to the top level of the payload
- updated tests

We're now passing the various UTM parameters from the raw page hit
through to the emitted request (to Tinybird). The change to move the UTM
parameters up was to maintain the purpose of the `parsedReferrer` field
which is more as an audit item for what the raw page hit passed to the
service (ie. what was parsed) to yield the various referrerUrl,
referrerMedium, and referrerSource fields in the result.