Added configurable `wait=true` to Tinybird events API calls (#368)
While testing the migration from JSON field to String field, I found
that Tinybird was accepting events but dropping them on the floor for a
split second during the deployment. It turns out that Tinybird's event
API will return `202: Accepted` to confirm receipt of the events, but
that doesn't guarantee the events are actually written to the database.
This means it's possible to get a `202: Accepted` from Tinybird, but the
events are potentially lost forever.

There is a `wait` query parameter that we can pass in our API requests,
which will wait for confirmation that the events were actually written
to the DB before returning. With this option set, the Tinybird API will
return an API error if the events are dropped for whatever reason (i.e.
during a deployment cutover). The worker process will receive the API
error and `nack` all the messages in the batch, which prompts them to be
retried.