Switched to using the server time for event timestamps (#459)
ref
https://linear.app/ghost/issue/NY-344/switch-to-using-server-time-for-analytics-events

We currently accept the timestamp in each event received, which is
generated from the client's browser, and forward it directly to
Tinybird. To avoid issues with browser clock drift, we've decided to
always use the server time, since we know it is right and can ensure it
is applied consistently. This is also simpler than validating the
timestamp and either rejecting or coercing the timestamp for "invalid"
timestamps.