Updated CI pipeline to require a single docker build for PR checks (testing & linting) and deploy after merge
ref NY-727
ref https://linear.app/ghost/issue/NY-727/eliminate-duplicate-docker-image-builds

During development of a PR we are continuously building a docker image to run testing and linting against. This makes a change so that each time we do this and tests and linting passes we push that image to GitHub Container Registry. Then, after merging the PR and starting a deploy, we first check if we can pull an image down that has already been built from this commit. If we can, great, deploy it. If we can't built, test and then deploy like normal.