Stopped running ci.yml workflow on push to main
This workflow is used to tell PRs if they are able to merge and has
already run on the code before the PR being merged. So it isn't needed
after the merge as well. Also, the same things (testing & linting) are
run, if needed, in the main workflow that does run after merge so it was
duplicate work.